Adding an 'editable' parameter to notify_* functions
The Pager created by all notify_* functions is editable by default,
which enables it to be selectable and take focus (which it does by
default unless 'editw' != 0 is passed) when pressing TAB.

This seems confusing to some users. Therefore, allow to make the
pager non-editable, so only buttons are selectable.